Update internal/handler/message_handler.go
The conversion 'uint64(orgID.(uint64))' is redundant since 'orgID' is already asserted as uint64. Consider simplifying the condition to 'if chat.OrganizationID != orgID.(uint64) {'.

+++ internal/handler/message_handler.go
@@ -157,7 +157,7 @@ func (h *MessageHandler) GetMessages(c *gin.Context) {
 	}
 
 	// Check if the chat belongs to the organization
-	if chat.OrganizationID != uint64(orgID.(uint64)) {
+	if chat.OrganizationID != orgID.(uint64) {
 		c.JSON(
 			http.StatusForbidden,
 			gin.H{"error": "You do not have permission to view messages in this chat"},
